refactor(providers): split permissions / geo / gallery providers out of app_providers.dart
Third batch of #4506. Moves three `@riverpod`-annotated providers to
`lib/providers/permissions_providers.dart`:

- geoBlockingService
- permissionsService
- gallerySaveService

The cluster has zero outbound dependencies on the rest of
`app_providers.dart`, so `app_providers.dart` only adds
`export 'permissions_providers.dart';` — no back-import needed.

Removed now-unused imports from `app_providers.dart`:
- `package:openvine/services/geo_blocking_service.dart`
- `package:openvine/services/gallery_save_service.dart`
- `package:permissions_service/permissions_service.dart`

Codegen moved: 3 `*Provider` chunks deleted from `app_providers.g.dart`,
matching chunks created in `permissions_providers.g.dart`.

The new file only needs `riverpod_annotation`; `flutter_riverpod` is
unnecessary because the cluster has no legacy `Provider<X>(...)` style
declarations.

Refs #4506, #4339 Wave 2
